+++ config/routes.rb
@@ -37,6 +37,7 @@ Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
       post "search", to: "searches#create"
+      post "therapists/:unique_id/messages", to: "therapist_messages#create"
     end
   end