+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_05_11_132500) do
+ActiveRecord::Schema[8.1].define(version: 2026_05_11_133100) do
   create_schema "extensions"
 
   # These are extensions that must be enabled in order to support this database